requirements.txt
fastapi 
pydantic
uvicorn
sqlalchemy 
asyncpg 
asyncio
passlib[argon2]
python-jose
python-multipart 
alembic
python-dotenv